+++ src/mjlab/tasks/velocity/config/rlboy/rl_cfg.py
@@ -48,4 +48,5 @@ def rlboy_ppo_runner_cfg() -> RslRlOnPolicyRunnerCfg:
     save_interval=50,
     num_steps_per_env=24,
     max_iterations=4_000,
+    clip_actions=1.0,
   )